pyproject.toml
[tool.poetry]
name = "python-poetry"
version = "1.2.0"
description = "A Python Docker image with Poetry installed and ready to use."
authors = ["Max Pfeiffer <max@maxpfeiffer.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.*"
click = "8.1.3"
docker = "6.1.3"
python-on-whales = "0.62.0"


[tool.poetry.dev-dependencies]
pylint = "2.17.4"
pytest = "7.3.1"
pytest-cov = "4.1.0"
coverage = "7.2.7"
black = "23.3.0"
pre-commit = "3.3.2"
semver = "3.0.0"
testcontainers = "3.7.1"
bcrypt = "4.0.1"
docker-image-py = "0.1.12"
python-slugify = "8.0.1"

# https://docs.pytest.org/en/latest/reference/customize.html
[tool.pytest.ini_options]
testpaths = [
    "tests",
]

# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
[tool.black]
line-length = 80
target-version = ["py39"]

[tool.pylint.main]
errors-only = true
recursive = "y"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
